requirements.txt
streamlit
pandas
pyyaml
openpyxl